requirements-lambda.txt
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.0.1
boto3==1.34.0 